+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+* Added `net.peer.name` and `net.peer.port` as dimensions on
+  `http.client.duration` metric.
+  ([#3907](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3907))
+
+* **Breaking change** `http.host` will no longer be populated on activity.
+  `net.peer.name` and `net.peer.port` attributes will be populated instead.
+  ([#3832](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3832))
+
+## 1.0.0-rc9.9
+
+Released 2022-Nov-07
+
 * Added back `netstandard2.0` target.
   ([#3787](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3787))